Implemented independent accept thread
Factored out accept() from second (communications) thread and placed
it in its own. When a new connection is received, a new GTBClient is
created and added to the acceptedQueue. SIGACCEPT is then pushed
to the comm thread where the connection is handled and the request
is received and added to the requestQueue.

A side-effect of this refactoring is the alteration to the parameters
handleConnection takes in. In order to make this change less
complicated, and because it served little purpose, handleConnection
no longer takes sockfd, the file descriptor of the server socket.