Solved race condition at startup
It only manifested when not running in GDB. Neither the Accept nor
the Comm thread blocked signals when they were spawned so when the
Accept thread signaled Comm that there is a client in the queue the
thread didn't know how to handle the signal even though it was
sigwaiting on it, so the program exited.

As a consequence, now on spawn, we block the signal until we call
sigwait in both threads because Comm now signals Accept when it is
ready.