Restart Comm thread correctly
When comm thread is restarted, make sure it starts handling pending
connections instead of waiting for a new connection to arrive first.

When the thread is restarted, reset lostcontrolat so that WDog won't
restart the thread again prematurely as a result of it not being
reset during the last cancel.