Avoid fatal segfault if we try to update empty PL
If we try to update an empty PatronList, we end up creating an
instance of PatronInfo without any field values. Not only is this
not a valid creation but it also segfaults on an IsInitialized()
check.

We attempt to avoid this situation entirely by verifying
that the number of PatronInfo instances we received is not zero
before we continue.