We authenticate via the webserver database
When we receive an AUTH request, we need to encode the necessary
data as a JSON and then encrypt it using GPG. Then encrypted file
is then sent to the webserver where it is decrypted and compared
against the users and signups in its database. If everything matches
then the webserver returns that the the users are authorized and the
information is stored in the local database for subsequent requests
(TODO). If they are not authorized then we reply as such (TODO).