If a ride is reassigned, return it on CURR
If a ride is reassigned, then the client needs to know about it and
remove the patron information from its list. We now accomplish this
by resending the patron information. The client should handle this by
detecting that it has already received this ride and removing the patron
from its list.